feat: configure the project
