+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+  "endOfLine": "auto",
+  "singleQuote": true,
+  "printWidth": 100
+}
